Moved the picking up of runfolders to...
... the processing units themselves. Feels better to keep them there.
